+++ src/client/AdminClient/src/folsAccessByUser.html
@@ -1,0 +1,76 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap" rel="stylesheet">
+
+    <link rel="stylesheet" href="https://fonts.sandbox.google.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,700,1,0"/>
+
+    <link rel="stylesheet" href="./css/global.css">
+    <link rel="stylesheet" href="./css/header.css">
+    <link rel="stylesheet" href="./css/footer.css">
+    <link rel="stylesheet" href="./css/folsAccess.css">
+
+    <title>FOLs Accesses</title>
+</head>
+<body>
+    <main class="root">
+        <header class="headerContainer">
+            <img src="./assets/images/header.png" class="headerImage">
+            <div class="headerContent">
+                <div class="menuButton" onmouseleave="global.closeMenu()">
+                    <span class="material-symbols-outlined icon" onclick="global.openMenu()">menu</span>
+                    <div class="menuOptions" id="menuOptions">
+                        <div class="menuOption" onclick="global.goToHome()">
+                            <span class="material-symbols-outlined">cottage</span>
+                            <label>Home</label>
+                        </div>
+                        <div class="menuOption">
+                            <span class="material-symbols-outlined">monitoring</span>
+                            <label>Graphics</label>
+                        </div>
+                        <div class="menuOption">
+                            <span class="material-symbols-outlined">text_snippet</span>
+                            <label>FOLs</label>
+                        </div>
+                    </div>
+                </div>
+                <label>FolConn</label>
+                <div class="exitButton" onclick="global.performLogout()">
+                    <span class="material-symbols-outlined icon">logout</span>
+                    <label>Log out</label>
+                </div>
+            </div>
+        </header>
+        <main class="mainContent">
+            <label>FOLs Accesses by users</label>
+            <div class="searchContainer">
+                <div class="searchFols container" id="searchFols">
+                    <div class="folSearchContainer" onclick="folsAccess.setFocusToInput('searchInput')">
+                        <input type="text" placeholder="Search FOLs by User" id="searchInput">
+                        <span class="material-symbols-outlined">search</span>
+                    </div>
+                    <div class="folsSearchResults" id="folSearchResults">
+
+                    </div>
+                </div>
+                <table class="searchResult container" id="searchResult">
+
+                </table>
+            </div>
+        </main>
+    </main>
+    <footer>
+        <img class="footer-image" src="./assets/images/footer.png"></img>
+        <div class="footer-content">
+            <h1 class="folcoon-footer">FolConn</h1>
+            <h1 class="copyright-footer">© 2022 Copyright - All rights reserved - FolConn</h1>
+        </div>
+    </footer>
+</body>
+</html>
