+++ src/client/AdminClient/src/home.html
@@ -51,6 +51,7 @@
             <label>What do you wish to see first?</label>
             <button class="folconnButton" onclick="home.goToApplicationAccesses()">Logins by regions</button>
             <button class="folconnButton" onclick="home.goToFolsAccesses()">FOLs accesses</button>
+            <button class="folconnButton" onclick="home.goToFolsAccessesByUser()">FOLs accesses by user</button>
         </div>
     </main>
     <footer>
