+++ src/client/AdminClient/src/ts/home.ts
@@ -5,3 +5,7 @@ export function goToApplicationAccesses(){
 export function goToFolsAccesses(){
     window.location.href = window.location.origin + "/fols-access";
 }
+
+export function goToFolsAccessesByUser(){
+    window.location.href = window.location.origin + "/fols-access-user";
+}
